[copilot] Add a flag --commits [<gitlogflags>] to output a commit log for the touched files.
Add a flag --commits [<gitlogflags>] that will change the output to a reverse-chronological list of commits that touched all of these files. Use git log [<gitlogflags>] to get that information and sort according to the date, which will appear in the log lines as 2025-02-21 14:23:07 -0800. Use pretty format "[ %h ] %ad | %<(20,trunc)%ae | %s" for the git log command.