[copilot] Create a README.md explaining the purpose of this script.
Add a README that describes the purpose of this script is to track ALL of the
commits that changed all of the files involved in a given feature (even zwhen
the commits were part of changes unrelated to the feature), where the feature
can be identified by a list of commits which can be created manually or
generated by e.g. git log --grep for feature keywords in commit messages. This
can be useful for identifying commits outside of explicit efforts on a feature
that might have affected the feature as a side effect.